[MIG] hr_payroll: Migration to 13.0
