+++ hr_payroll/data/hr_payroll_data.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <odoo>
-    <data noupdate="0">
+    <data noupdate="1">
         <record id="contrib_register_employees" model="hr.contribution.register">
             <field name="name">Employees</field>
             <field name="partner_id" eval="False" />
