+++ hr_payroll/models/__init__.py
@@ -1,7 +1,16 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from . import hr_contract
+from . import hr_contract_advantage_template
 from . import hr_employee
-from . import res_config_settings
+from . import hr_payroll_structure
 from . import hr_salary_rule
+from . import hr_salary_rule_category
+from . import hr_rule_input
+from . import hr_contribution_register
 from . import hr_payslip
+from . import hr_payslip_line
+from . import hr_payslip_input
+from . import hr_payslip_worked_days
+from . import hr_payslip_run
+from . import res_config_settings
