+++ hr_payroll/models/hr_contract.py
@@ -1,6 +1,6 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from odoo import api, fields, models
+from odoo import fields, models
 
 
 class HrContract(models.Model):
@@ -32,11 +32,11 @@ class HrContract(models.Model):
         required=True, help="Employee's working schedule."
     )
 
-    @api.multi
     def get_all_structures(self):
         """
-        @return: the structures linked to the given contracts, ordered by hierachy (parent=False first,
-                 then first level children and so on) and without duplicata
+        @return: the structures linked to the given contracts, ordered by
+                 hierachy (parent=False first, then first level children and
+                 so on) and without duplicata
         """
         structures = self.mapped("struct_id")
         if not structures:
@@ -44,13 +44,11 @@ class HrContract(models.Model):
         # YTI TODO return browse records
         return list(set(structures._get_parent_structure().ids))
 
-    @api.multi
     def get_attribute(self, code, attribute):
         return self.env["hr.contract.advantage.template"].search(
             [("code", "=", code)], limit=1
         )[attribute]
 
-    @api.multi
     def set_attribute_value(self, code, active):
         for contract in self:
             if active:
@@ -62,18 +60,3 @@ class HrContract(models.Model):
                 contract[code] = value
             else:
                 contract[code] = 0.0
-
-
-class HrContractAdvandageTemplate(models.Model):
-    _name = "hr.contract.advantage.template"
-    _description = "Employee's Advantage on Contract"
-
-    name = fields.Char("Name", required=True)
-    code = fields.Char("Code", required=True)
-    lower_bound = fields.Float(
-        "Lower Bound", help="Lower bound authorized by the employer for this advantage"
-    )
-    upper_bound = fields.Float(
-        "Upper Bound", help="Upper bound authorized by the employer for this advantage"
-    )
-    default_value = fields.Float("Default value for this advantage")
