+++ hr_payroll/models/hr_employee.py
@@ -1,6 +1,6 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from odoo import api, fields, models
+from odoo import fields, models
 
 
 class HrEmployee(models.Model):
@@ -16,7 +16,6 @@ class HrEmployee(models.Model):
         groups="hr_payroll.group_hr_payroll_user",
     )
 
-    @api.multi
     def _compute_payslip_count(self):
         for employee in self:
             employee.payslip_count = len(employee.slip_ids)
