+++ hr_payroll/models/hr_payslip.py
@@ -9,7 +9,83 @@ from pytz import timezone
 from odoo import _, api, fields, models, tools
 from odoo.exceptions import UserError, ValidationError
 
-from odoo.addons import decimal_precision as dp
+
+# These classes are used in the _get_payslip_lines() method
+class BrowsableObject(object):
+    def __init__(self, employee_id, vals_dict, env):
+        self.employee_id = employee_id
+        self.dict = vals_dict
+        self.env = env
+
+    def __getattr__(self, attr):
+        return attr in self.dict and self.dict.__getitem__(attr) or 0.0
+
+
+class InputLine(BrowsableObject):
+    """a class that will be used into the python code, mainly for
+    usability purposes"""
+
+    def sum(self, code, from_date, to_date=None):
+        if to_date is None:
+            to_date = fields.Date.today()
+        self.env.cr.execute(
+            """
+            SELECT sum(amount) as sum
+            FROM hr_payslip as hp, hr_payslip_input as pi
+            WHERE hp.employee_id = %s AND hp.state = 'done'
+            AND hp.date_from >= %s AND hp.date_to <= %s
+            AND hp.id = pi.payslip_id AND pi.code = %s""",
+            (self.employee_id, from_date, to_date, code),
+        )
+        return self.env.cr.fetchone()[0] or 0.0
+
+
+class WorkedDays(BrowsableObject):
+    """a class that will be used into the python code, mainly for
+     usability purposes"""
+
+    def _sum(self, code, from_date, to_date=None):
+        if to_date is None:
+            to_date = fields.Date.today()
+        self.env.cr.execute(
+            """
+            SELECT sum(number_of_days) as number_of_days,
+             sum(number_of_hours) as number_of_hours
+            FROM hr_payslip as hp, hr_payslip_worked_days as pi
+            WHERE hp.employee_id = %s AND hp.state = 'done'
+            AND hp.date_from >= %s AND hp.date_to <= %s
+            AND hp.id = pi.payslip_id AND pi.code = %s""",
+            (self.employee_id, from_date, to_date, code),
+        )
+        return self.env.cr.fetchone()
+
+    def sum(self, code, from_date, to_date=None):
+        res = self._sum(code, from_date, to_date)
+        return res and res[0] or 0.0
+
+    def sum_hours(self, code, from_date, to_date=None):
+        res = self._sum(code, from_date, to_date)
+        return res and res[1] or 0.0
+
+
+class Payslips(BrowsableObject):
+    """a class that will be used into the python code, mainly for
+     usability purposes"""
+
+    def sum(self, code, from_date, to_date=None):
+        if to_date is None:
+            to_date = fields.Date.today()
+        self.env.cr.execute(
+            """SELECT sum(case when hp.credit_note = False then
+            (pl.total) else (-pl.total) end)
+                    FROM hr_payslip as hp, hr_payslip_line as pl
+                    WHERE hp.employee_id = %s AND hp.state = 'done'
+                    AND hp.date_from >= %s AND hp.date_to <= %s AND
+                     hp.id = pl.slip_id AND pl.code = %s""",
+            (self.employee_id, from_date, to_date, code),
+        )
+        res = self.env.cr.fetchone()
+        return res and res[0] or 0.0
 
 
 class HrPayslip(models.Model):
@@ -21,10 +97,11 @@ class HrPayslip(models.Model):
         string="Structure",
         readonly=True,
         states={"draft": [("readonly", False)]},
-        help="Defines the rules that have to be applied to this payslip, accordingly "
-        "to the contract chosen. If you let empty the field contract, this field isn't "
-        "mandatory anymore and thus the rules applied will be all the rules set on the "
-        "structure of all contracts of the employee valid for the chosen period",
+        help="Defines the rules that have to be applied to this payslip, "
+        "accordingly to the contract chosen. If you let empty the field "
+        "contract, this field isn't mandatory anymore and thus the rules "
+        "applied will be all the rules set on the structure of all contracts "
+        "of the employee valid for the chosen period",
     )
     name = fields.Char(
         string="Payslip Name", readonly=True, states={"draft": [("readonly", False)]}
@@ -58,7 +135,8 @@ class HrPayslip(models.Model):
         ),
         states={"draft": [("readonly", False)]},
     )
-    # this is chaos: 4 states are defined, 3 are used ('verify' isn't) and 5 exist ('confirm' seems to have existed)
+    # this is chaos: 4 states are defined, 3 are used ('verify' isn't) and 5
+    # exist ('confirm' seems to have existed)
     state = fields.Selection(
         [
             ("draft", "Draft"),
@@ -72,9 +150,9 @@ class HrPayslip(models.Model):
         copy=False,
         default="draft",
         help="""* When the payslip is created the status is \'Draft\'
-                \n* If the payslip is under verification, the status is \'Waiting\'.
-                \n* If the payslip is confirmed then status is set to \'Done\'.
-                \n* When user cancel payslip the status is \'Rejected\'.""",
+        \n* If the payslip is under verification, the status is \'Waiting\'.
+        \n* If the payslip is confirmed then status is set to \'Done\'.
+        \n* When user cancel payslip the status is \'Rejected\'.""",
     )
     line_ids = fields.One2many(
         "hr.payslip.line",
@@ -88,7 +166,7 @@ class HrPayslip(models.Model):
         string="Company",
         readonly=True,
         copy=False,
-        default=lambda self: self.env["res.company"]._company_default_get(),
+        default=lambda self: self.env.company,
         states={"draft": [("readonly", False)]},
     )
     worked_days_line_ids = fields.One2many(
@@ -143,14 +221,12 @@ class HrPayslip(models.Model):
         compute="_compute_payslip_count", string="Payslip Computation Details"
     )
 
-    @api.multi
     def _compute_details_by_salary_rule_category(self):
         for payslip in self:
             payslip.details_by_salary_rule_category = payslip.mapped(
                 "line_ids"
             ).filtered(lambda line: line.category_id)
 
-    @api.multi
     def _compute_payslip_count(self):
         for payslip in self:
             payslip.payslip_count = len(payslip.line_ids)
@@ -158,39 +234,35 @@ class HrPayslip(models.Model):
     @api.constrains("date_from", "date_to")
     def _check_dates(self):
         if any(self.filtered(lambda payslip: payslip.date_from > payslip.date_to)):
-            raise ValidationError(_("Payslip 'Date From' must be earlier 'Date To'."))
+            raise ValidationError(
+                _("Payslip 'Date From' must be earlier than 'Date To'.")
+            )
 
-    @api.multi
-    @api.returns("self", lambda value: value.id)
     def copy(self, default=None):
-        rec = super(HrPayslip, self).copy(default)
+        rec = super().copy(default)
         for l in self.input_line_ids:
             l.copy({"payslip_id": rec.id})
         for l in self.line_ids:
             l.copy({"slip_id": rec.id, "input_ids": []})
         return rec
 
-    @api.multi
     def action_payslip_draft(self):
         return self.write({"state": "draft"})
 
-    @api.multi
     def action_payslip_done(self):
         if not self.env.context.get("without_compute_sheet"):
             self.compute_sheet()
         return self.write({"state": "done"})
 
-    @api.multi
     def action_payslip_cancel(self):
         if self.filtered(lambda slip: slip.state == "done"):
             raise UserError(_("Cannot cancel a payslip that is done."))
         return self.write({"state": "cancel"})
 
-    @api.multi
     def refund_sheet(self):
         for payslip in self:
             copied_payslip = payslip.copy(
-                {"credit_note": True, "name": _("Refund: ") + payslip.name}
+                {"credit_note": True, "name": _("Refund: %s") % payslip.name}
             )
             number = copied_payslip.number or self.env["ir.sequence"].next_by_code(
                 "salary.slip"
@@ -199,13 +271,12 @@ class HrPayslip(models.Model):
             copied_payslip.with_context(
                 without_compute_sheet=True
             ).action_payslip_done()
-        formview_ref = self.env.ref("hr_payroll.view_hr_payslip_form", False)
-        treeview_ref = self.env.ref("hr_payroll.view_hr_payslip_tree", False)
+        formview_ref = self.env.ref("hr_payroll.hr_payslip_view_form", False)
+        treeview_ref = self.env.ref("hr_payroll.hr_payslip_view_tree", False)
         return {
             "name": ("Refund Payslip"),
             "view_mode": "tree, form",
             "view_id": False,
-            "view_type": "form",
             "res_model": "hr.payslip",
             "type": "ir.actions.act_window",
             "target": "current",
@@ -217,15 +288,13 @@ class HrPayslip(models.Model):
             "context": {},
         }
 
-    @api.multi
     def check_done(self):
         return True
 
-    @api.multi
     def unlink(self):
         if any(self.filtered(lambda payslip: payslip.state not in ("draft", "cancel"))):
             raise UserError(
-                _("You cannot delete a payslip which is not draft or cancelled!")
+                _("You cannot delete a payslip which is not draft or cancelled")
             )
         return super(HrPayslip, self).unlink()
 
@@ -236,13 +305,15 @@ class HrPayslip(models.Model):
         @param employee: recordset of employee
         @param date_from: date field
         @param date_to: date field
-        @return: returns the ids of all the contracts for the given employee that need to be considered for the given dates
+        @return: returns the ids of all the contracts for the given employee
+        that need to be considered for the given dates
         """
         # a contract is valid if it ends between the given dates
         clause_1 = ["&", ("date_end", "<=", date_to), ("date_end", ">=", date_from)]
         # OR if it starts between the given dates
         clause_2 = ["&", ("date_start", "<=", date_to), ("date_start", ">=", date_from)]
-        # OR if it starts before the date_from and finish after the date_end (or never finish)
+        # OR if it starts before the date_from and finish after the date_end
+        # (or never finish)
         clause_3 = [
             "&",
             ("date_start", "<=", date_from),
@@ -258,7 +329,6 @@ class HrPayslip(models.Model):
         )
         return self.env["hr.contract"].search(clause_final).ids
 
-    @api.multi
     def compute_sheet(self):
         for payslip in self:
             number = payslip.number or self.env["ir.sequence"].next_by_code(
@@ -267,7 +337,8 @@ class HrPayslip(models.Model):
             # delete old payslip lines
             payslip.line_ids.unlink()
             # set the list of contract for which the rules have to be applied
-            # if we don't give the contract, then the rules to apply should be for all current contracts of the employee
+            # if we don't give the contract, then the rules to apply should be
+            # for all current contracts of the employee
             contract_ids = payslip.contract_id.ids or self.get_contract(
                 payslip.employee_id, payslip.date_from, payslip.date_to
             )
@@ -282,15 +353,16 @@ class HrPayslip(models.Model):
     def get_worked_day_lines(self, contracts, date_from, date_to):
         """
         @param contract: Browse record of contracts
-        @return: returns a list of dict containing the input that should be applied for the given contract between date_from and date_to
+        @return: returns a list of dict containing the input that should be
+        applied for the given contract between date_from and date_to
         """
         res = []
         # fill only if the contract as a working schedule linked
         for contract in contracts.filtered(
             lambda contract: contract.resource_calendar_id
         ):
-            day_from = datetime.combine(fields.Date.from_string(date_from), time.min)
-            day_to = datetime.combine(fields.Date.from_string(date_to), time.max)
+            day_from = datetime.combine(date_from, time.min)
+            day_to = datetime.combine(date_to, time.max)
 
             # compute leave days
             leaves = {}
@@ -322,7 +394,7 @@ class HrPayslip(models.Model):
                     current_leave_struct["number_of_days"] += hours / work_hours
 
             # compute worked days
-            work_data = contract.employee_id.get_work_days_data(
+            work_data = contract.employee_id._get_work_days_data(
                 day_from, day_to, calendar=contract.resource_calendar_id
             )
             attendances = {
@@ -347,16 +419,19 @@ class HrPayslip(models.Model):
             self.env["hr.payroll.structure"].browse(structure_ids).get_all_rules()
         )
         sorted_rule_ids = [id for id, sequence in sorted(rule_ids, key=lambda x: x[1])]
-        inputs = self.env["hr.salary.rule"].browse(sorted_rule_ids).mapped("input_ids")
+        payslip_inputs = (
+            self.env["hr.salary.rule"].browse(sorted_rule_ids).mapped("input_ids")
+        )
 
         for contract in contracts:
-            for input in inputs:
-                input_data = {
-                    "name": input.name,
-                    "code": input.code,
-                    "contract_id": contract.id,
-                }
-                res += [input_data]
+            for payslip_input in payslip_inputs:
+                res.append(
+                    {
+                        "name": payslip_input.name,
+                        "code": payslip_input.code,
+                        "contract_id": contract.id,
+                    }
+                )
         return res
 
     @api.model
@@ -374,72 +449,8 @@ class HrPayslip(models.Model):
 
             return localdict
 
-        class BrowsableObject(object):
-            def __init__(self, employee_id, dict, env):
-                self.employee_id = employee_id
-                self.dict = dict
-                self.env = env
-
-            def __getattr__(self, attr):
-                return attr in self.dict and self.dict.__getitem__(attr) or 0.0
-
-        class InputLine(BrowsableObject):
-            """a class that will be used into the python code, mainly for usability purposes"""
-
-            def sum(self, code, from_date, to_date=None):
-                if to_date is None:
-                    to_date = fields.Date.today()
-                self.env.cr.execute(
-                    """
-                    SELECT sum(amount) as sum
-                    FROM hr_payslip as hp, hr_payslip_input as pi
-                    WHERE hp.employee_id = %s AND hp.state = 'done'
-                    AND hp.date_from >= %s AND hp.date_to <= %s AND hp.id = pi.payslip_id AND pi.code = %s""",
-                    (self.employee_id, from_date, to_date, code),
-                )
-                return self.env.cr.fetchone()[0] or 0.0
-
-        class WorkedDays(BrowsableObject):
-            """a class that will be used into the python code, mainly for usability purposes"""
-
-            def _sum(self, code, from_date, to_date=None):
-                if to_date is None:
-                    to_date = fields.Date.today()
-                self.env.cr.execute(
-                    """
-                    SELECT sum(number_of_days) as number_of_days, sum(number_of_hours) as number_of_hours
-                    FROM hr_payslip as hp, hr_payslip_worked_days as pi
-                    WHERE hp.employee_id = %s AND hp.state = 'done'
-                    AND hp.date_from >= %s AND hp.date_to <= %s AND hp.id = pi.payslip_id AND pi.code = %s""",
-                    (self.employee_id, from_date, to_date, code),
-                )
-                return self.env.cr.fetchone()
-
-            def sum(self, code, from_date, to_date=None):
-                res = self._sum(code, from_date, to_date)
-                return res and res[0] or 0.0
-
-            def sum_hours(self, code, from_date, to_date=None):
-                res = self._sum(code, from_date, to_date)
-                return res and res[1] or 0.0
-
-        class Payslips(BrowsableObject):
-            """a class that will be used into the python code, mainly for usability purposes"""
-
-            def sum(self, code, from_date, to_date=None):
-                if to_date is None:
-                    to_date = fields.Date.today()
-                self.env.cr.execute(
-                    """SELECT sum(case when hp.credit_note = False then (pl.total) else (-pl.total) end)
-                            FROM hr_payslip as hp, hr_payslip_line as pl
-                            WHERE hp.employee_id = %s AND hp.state = 'done'
-                            AND hp.date_from >= %s AND hp.date_to <= %s AND hp.id = pl.slip_id AND pl.code = %s""",
-                    (self.employee_id, from_date, to_date, code),
-                )
-                res = self.env.cr.fetchone()
-                return res and res[0] or 0.0
-
-        # we keep a dict with the result because a value can be overwritten by another rule with the same code
+        # we keep a dict with the result because a value can be overwritten by
+        # another rule with the same code
         result_dict = {}
         rules_dict = {}
         worked_days_dict = {}
@@ -464,7 +475,8 @@ class HrPayslip(models.Model):
             "worked_days": worked_days,
             "inputs": inputs,
         }
-        # get the ids of the structures on the contracts and their parent id as well
+        # get the ids of the structures on the contracts and their parent id
+        # as well
         contracts = self.env["hr.contract"].browse(contract_ids)
         if len(contracts) == 1 and payslip.struct_id:
             structure_ids = list(set(payslip.struct_id._get_parent_structure().ids))
@@ -494,7 +506,8 @@ class HrPayslip(models.Model):
                     previous_amount = (
                         rule.code in localdict and localdict[rule.code] or 0.0
                     )
-                    # set/overwrite the amount computed for this rule in the localdict
+                    # set/overwrite the amount computed for this rule in the
+                    # localdict
                     tot_rule = amount * qty * rate / 100.0
                     localdict[rule.code] = tot_rule
                     rules_dict[rule.code] = rule
@@ -533,10 +546,7 @@ class HrPayslip(models.Model):
 
         return list(result_dict.values())
 
-    # YTI TODO To rename. This method is not really an onchange, as it is not in any view
-    # employee_id and contract_id could be browse records
-    @api.multi
-    def onchange_employee_id(
+    def get_payslip_vals(
         self, date_from, date_to, employee_id=False, contract_id=False
     ):
         # defaults
@@ -544,12 +554,10 @@ class HrPayslip(models.Model):
             "value": {
                 "line_ids": [],
                 # delete old input lines
-                "input_line_ids": [(2, x,) for x in self.input_line_ids.ids],
+                "input_line_ids": [(2, x) for x in self.input_line_ids.ids],
                 # delete old worked days lines
-                "worked_days_line_ids": [
-                    (2, x,) for x in self.worked_days_line_ids.ids
-                ],
-                #'details_by_salary_head':[], TODO put me back
+                "worked_days_line_ids": [(2, x) for x in self.worked_days_line_ids.ids],
+                # 'details_by_salary_head':[], TODO put me back
                 "name": "",
                 "contract_id": False,
                 "struct_id": False,
@@ -557,7 +565,7 @@ class HrPayslip(models.Model):
         }
         if (not employee_id) or (not date_from) or (not date_to):
             return res
-        ttyme = datetime.combine(fields.Date.from_string(date_from), time.min)
+        ttyme = datetime.combine(date_from, time.min)
         employee = self.env["hr.employee"].browse(employee_id)
         locale = self.env.context.get("lang") or "en_US"
         res["value"].update(
@@ -583,7 +591,8 @@ class HrPayslip(models.Model):
                 # set the list of contract for which the input have to be filled
                 contract_ids = [contract_id]
             else:
-                # if we don't give the contract, then the input to fill should be for all current contracts of the employee
+                # if we don't give the contract, then the input to fill should
+                # be for all current contracts of the employee
                 contract_ids = self.get_contract(employee, date_from, date_to)
 
         if not contract_ids:
@@ -593,9 +602,7 @@ class HrPayslip(models.Model):
         struct = contract.struct_id
         if not struct:
             return res
-        res["value"].update(
-            {"struct_id": struct.id,}
-        )
+        res["value"].update({"struct_id": struct.id})
         # computation of the salary input
         contracts = self.env["hr.contract"].browse(contract_ids)
         worked_days_line_ids = self.get_worked_day_lines(contracts, date_from, date_to)
@@ -617,9 +624,9 @@ class HrPayslip(models.Model):
         employee = self.employee_id
         date_from = self.date_from
         date_to = self.date_to
-        contract_ids = []
+        contract_ids = self.contract_id.ids
 
-        ttyme = datetime.combine(fields.Date.from_string(date_from), time.min)
+        ttyme = datetime.combine(date_from, time.min)
         locale = self.env.context.get("lang") or "en_US"
         self.name = _("Salary Slip of %s for %s") % (
             employee.name,
@@ -668,157 +675,3 @@ class HrPayslip(models.Model):
             return line[0].total
         else:
             return 0.0
-
-
-class HrPayslipLine(models.Model):
-    _name = "hr.payslip.line"
-    _inherit = "hr.salary.rule"
-    _description = "Payslip Line"
-    _order = "contract_id, sequence"
-
-    slip_id = fields.Many2one(
-        "hr.payslip", string="Pay Slip", required=True, ondelete="cascade"
-    )
-    salary_rule_id = fields.Many2one("hr.salary.rule", string="Rule", required=True)
-    employee_id = fields.Many2one("hr.employee", string="Employee", required=True)
-    contract_id = fields.Many2one(
-        "hr.contract", string="Contract", required=True, index=True
-    )
-    rate = fields.Float(
-        string="Rate (%)", digits=dp.get_precision("Payroll Rate"), default=100.0
-    )
-    amount = fields.Float(digits=dp.get_precision("Payroll"))
-    quantity = fields.Float(digits=dp.get_precision("Payroll"), default=1.0)
-    total = fields.Float(
-        compute="_compute_total",
-        string="Total",
-        digits=dp.get_precision("Payroll"),
-        store=True,
-    )
-
-    @api.depends("quantity", "amount", "rate")
-    def _compute_total(self):
-        for line in self:
-            line.total = float(line.quantity) * line.amount * line.rate / 100
-
-    @api.model_create_multi
-    def create(self, vals_list):
-        for values in vals_list:
-            if "employee_id" not in values or "contract_id" not in values:
-                payslip = self.env["hr.payslip"].browse(values.get("slip_id"))
-                values["employee_id"] = (
-                    values.get("employee_id") or payslip.employee_id.id
-                )
-                values["contract_id"] = (
-                    values.get("contract_id")
-                    or payslip.contract_id
-                    and payslip.contract_id.id
-                )
-                if not values["contract_id"]:
-                    raise UserError(
-                        _("You must set a contract to create a payslip line.")
-                    )
-        return super(HrPayslipLine, self).create(vals_list)
-
-
-class HrPayslipWorkedDays(models.Model):
-    _name = "hr.payslip.worked_days"
-    _description = "Payslip Worked Days"
-    _order = "payslip_id, sequence"
-
-    name = fields.Char(string="Description", required=True)
-    payslip_id = fields.Many2one(
-        "hr.payslip", string="Pay Slip", required=True, ondelete="cascade", index=True
-    )
-    sequence = fields.Integer(required=True, index=True, default=10)
-    code = fields.Char(
-        required=True, help="The code that can be used in the salary rules"
-    )
-    number_of_days = fields.Float(string="Number of Days")
-    number_of_hours = fields.Float(string="Number of Hours")
-    contract_id = fields.Many2one(
-        "hr.contract",
-        string="Contract",
-        required=True,
-        help="The contract for which applied this input",
-    )
-
-
-class HrPayslipInput(models.Model):
-    _name = "hr.payslip.input"
-    _description = "Payslip Input"
-    _order = "payslip_id, sequence"
-
-    name = fields.Char(string="Description", required=True)
-    payslip_id = fields.Many2one(
-        "hr.payslip", string="Pay Slip", required=True, ondelete="cascade", index=True
-    )
-    sequence = fields.Integer(required=True, index=True, default=10)
-    code = fields.Char(
-        required=True, help="The code that can be used in the salary rules"
-    )
-    amount = fields.Float(
-        help="It is used in computation. For e.g. A rule for sales having "
-        "1% commission of basic salary for per product can defined in expression "
-        "like result = inputs.SALEURO.amount * contract.wage*0.01."
-    )
-    contract_id = fields.Many2one(
-        "hr.contract",
-        string="Contract",
-        required=True,
-        help="The contract for which applied this input",
-    )
-
-
-class HrPayslipRun(models.Model):
-    _name = "hr.payslip.run"
-    _description = "Payslip Batches"
-
-    name = fields.Char(
-        required=True, readonly=True, states={"draft": [("readonly", False)]}
-    )
-    slip_ids = fields.One2many(
-        "hr.payslip",
-        "payslip_run_id",
-        string="Payslips",
-        readonly=True,
-        states={"draft": [("readonly", False)]},
-    )
-    state = fields.Selection(
-        [("draft", "Draft"), ("close", "Close"),],
-        string="Status",
-        index=True,
-        readonly=True,
-        copy=False,
-        default="draft",
-    )
-    date_start = fields.Date(
-        string="Date From",
-        required=True,
-        readonly=True,
-        states={"draft": [("readonly", False)]},
-        default=lambda self: fields.Date.to_string(date.today().replace(day=1)),
-    )
-    date_end = fields.Date(
-        string="Date To",
-        required=True,
-        readonly=True,
-        states={"draft": [("readonly", False)]},
-        default=lambda self: fields.Date.to_string(
-            (datetime.now() + relativedelta(months=+1, day=1, days=-1)).date()
-        ),
-    )
-    credit_note = fields.Boolean(
-        string="Credit Note",
-        readonly=True,
-        states={"draft": [("readonly", False)]},
-        help="If its checked, indicates that all payslips generated from here are refund payslips.",
-    )
-
-    @api.multi
-    def draft_payslip_run(self):
-        return self.write({"state": "draft"})
-
-    @api.multi
-    def close_payslip_run(self):
-        return self.write({"state": "close"})
