+++ hr_payroll/readme/CONTRIBUTORS.rst
@@ -1,0 +1,2 @@
+* Odoo SA <info@odoo.com>
+* David James <david@djdc.net.au>
