+++ hr_payroll/readme/DESCRIPTION.rst
@@ -1,0 +1,3 @@
+Manage your employee payroll records.
+
+This module is a backport from Odoo SA and as such, it is not included in the OCA CLA. That means we do not have a copy of the copyright on it like all other OCA modules.
