+++ hr_payroll/report/report.xml
