+++ hr_payroll/views/hr_contract_advantage_views.xml
@@ -1,0 +1,53 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo>
+    <record id="hr_contract_advantage_template_view_form" model="ir.ui.view">
+        <field name="name">hr.contract.advantage.template.form</field>
+        <field name="model">hr.contract.advantage.template</field>
+        <field name="arch" type="xml">
+            <form>
+                <sheet>
+                    <div class="oe_title">
+                        <h1>
+                            <field name="name" placeholder="Advantage Name" />
+                        </h1>
+                    </div>
+                    <group>
+                        <group>
+                            <field name="code" />
+                        </group>
+                    </group>
+                    <group>
+                        <group>
+                            <field name="default_value" />
+                            <field name="lower_bound" />
+                            <field name="upper_bound" />
+                        </group>
+                    </group>
+                </sheet>
+            </form>
+        </field>
+    </record>
+    <record id="hr_contract_advantage_template_view_tree" model="ir.ui.view">
+        <field name="name">hr.contract.advantage.template.tree</field>
+        <field name="model">hr.contract.advantage.template</field>
+        <field name="arch" type="xml">
+            <tree>
+                <field name="name" />
+                <field name="code" />
+                <field name="lower_bound" />
+                <field name="upper_bound" />
+                <field name="default_value" />
+            </tree>
+        </field>
+    </record>
+    <record id="hr_contract_advantage_template_action" model="ir.actions.act_window">
+        <field name="name">Contract Advantage Templates</field>
+        <field name="res_model">hr.contract.advantage.template</field>
+    </record>
+    <menuitem
+        id="hr_contract_advantage_template_menu_action"
+        action="hr_contract_advantage_template_action"
+        parent="hr_payroll_menu_configuration"
+        sequence="50"
+    />
+</odoo>
