+++ hr_payroll/views/hr_employee_views.xml
@@ -1,14 +1,14 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <odoo>
     <!-- Employee View -->
-    <record id="payroll_hr_employee_view_form" model="ir.ui.view">
-        <field name="name">payroll.hr.employee.view.form</field>
+    <record id="hr_employee_view_form" model="ir.ui.view">
+        <field name="name">hr.employee.view.form.payroll</field>
         <field name="model">hr.employee</field>
         <field name="inherit_id" ref="hr.view_employee_form" />
         <field name="arch" type="xml">
-            <button name="toggle_active" position="before">
+            <xpath expr="//div[@name='button_box']" position="inside">
                 <button
-                    name="%(act_hr_employee_payslip_list)d"
+                    name="%(hr_payslip_action_employee)d"
                     class="oe_stat_button"
                     icon="fa-money"
                     type="action"
@@ -16,7 +16,7 @@
                 >
                     <field name="payslip_count" widget="statinfo" string="Payslips" />
                 </button>
-            </button>
+            </xpath>
         </field>
     </record>
 </odoo>
