+++ hr_payroll/views/report_contributionregister.xml
@@ -48,13 +48,13 @@
                                     <td class="text-right">
                                         <span
                                             t-esc="line.amount"
-                                            t-options='{"widget": "monetary", "display_currency": o.company_id.currency_id}'
+                                            t-esc-options='{"widget": "monetary", "display_currency": o.company_id.currency_id}'
                                         />
                                     </td>
                                     <td class="text-right">
                                         <span
                                             t-esc="line.total"
-                                            t-options='{"widget": "monetary", "display_currency": o.company_id.currency_id}'
+                                            t-esc-options='{"widget": "monetary", "display_currency": o.company_id.currency_id}'
                                         />
                                     </td>
                                 </tr>
@@ -70,7 +70,7 @@
                                         <td class="text-right">
                                             <span
                                                 t-esc="lines_total.get(o.id)"
-                                                t-options='{"widget": "monetary", "display_currency": o.company_id.currency_id}'
+                                                t-esc-options='{"widget": "monetary", "display_currency": o.company_id.currency_id}'
                                             />
                                         </td>
                                     </tr>
