+++ hr_payroll/views/report_payslipdetails.xml
@@ -100,7 +100,7 @@
                                     <td class="text-right">
                                         <span
                                             t-esc="h['total']"
-                                            t-options='{"widget": "monetary", "display_currency": o.company_id.currency_id}'
+                                            t-esc-options='{"widget": "monetary", "display_currency": o.company_id.currency_id}'
                                         />
                                     </td>
                                 </tr>
@@ -137,7 +137,7 @@
                                     <td class="text-right">
                                         <span
                                             t-esc="p.get('total', 0)"
-                                            t-options='{"widget": "monetary", "display_currency": o.company_id.currency_id}'
+                                            t-esc-options='{"widget": "monetary", "display_currency": o.company_id.currency_id}'
                                         />
                                     </td>
                                 </tr>
