+++ hr_payroll/wizard/hr_payroll_contribution_register_report_views.xml
@@ -29,7 +29,6 @@
         <field name="name">PaySlip Lines</field>
         <field name="type">ir.actions.act_window</field>
         <field name="res_model">payslip.lines.contribution.register</field>
-        <field name="view_type">form</field>
         <field name="view_mode">form</field>
         <field name="target">new</field>
         <field name="binding_model_id" ref="model_hr_contribution_register" />
