+++ hr_payroll/wizard/hr_payroll_payslips_by_employees_views.xml
@@ -31,7 +31,6 @@
     <record id="action_hr_payslip_by_employees" model="ir.actions.act_window">
         <field name="name">Generate Payslips</field>
         <field name="res_model">hr.payslip.employees</field>
-        <field name="view_type">form</field>
         <field name="view_mode">tree,form</field>
         <field name="view_id" ref="view_hr_payslip_by_employees" />
         <field name="target">new</field>
