+++ setup/hr_payroll/odoo/addons/hr_payroll
@@ -1,0 +1,1 @@
+../../../../hr_payroll
